testing things first with one channel, i'll expand later on it

+++ arduino/utils/analog.h
@@ -2,6 +2,6 @@
 #define _ANALOG_H_
 
 void ADC_init(void);
-uint16_t ADC_read(uint8_t channel);
+uint8_t ADC_read(uint8_t channel);
 
 #endif
